backend/requirements.txt
# ─── Web Framework ─────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ─── Data Validation ───────────────────────────────────────────────────────
pydantic==2.8.2

# ─── Auth / Security ───────────────────────────────────────────────────────
passlib[pbkdf2-sha256]==1.7.4

# ─── HTTP Client ───────────────────────────────────────────────────────────
requests>=2.32.0
httpx>=0.27.0

# ─── Database — PostgreSQL only (SQLite removed) ───────────────────────────
psycopg2-binary>=2.9.9

# ─── RAG / Embeddings ──────────────────────────────────────────────────────
sentence-transformers>=3.0.0

# ─── Document Parsing ──────────────────────────────────────────────────────
python-docx>=1.1.0
pypdf>=4.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# ─── Data Processing ───────────────────────────────────────────────────────
pandas>=2.2.0
openpyxl>=3.1.0

# ─── Reporting / Export ────────────────────────────────────────────────────
reportlab>=4.0.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2

# ─── Utilities ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
